cn-predictor/requirements.txt
numpy==1.26.4
pandas
scikit-learn
rdkit-pypi
joblib
streamlit
io